Create README.md
This is a game with a paddle you move to hit a ball. The ball bounces up to hit bricks. Destroying the bricks gains points.
You win the game when all the bricks are destroyed by the ball. 
This program works in Pycharm and VSCode.
This was for my 100 Days of Code class.